Implemented Waypoint, Teleporter and IntegrityMonitor

+++ src/main/scala/me/amuxix/util/Entity.scala
@@ -10,6 +10,6 @@ object Entity {
 }
 abstract class Entity {
   def location: Option[Location]
-  def pitch: Double
-  def yaw: Double
+  def pitch: Float
+  def yaw: Float
 }
